splitting DTCInterface into a common and a trk specific part
