+++ otsdaq-mu2e-tracker/Ui/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 cet_make_library(
         SOURCE          
-        DtcInterface.cc DtcInterface_print.cc DtcInterface_ControlRoc.cc CfoInterface.cc
+        DtcInterface.cc DtcInterfaceBase.cc DtcInterface_print.cc DtcInterface_ControlRoc.cc CfoInterface.cc
         LIBRARIES PUBLIC
 	otsdaq_mu2e_tracker::otsdaq-mu2e-tracker_ParseAlignment
         artdaq::DAQdata
